+++ CHANGES.rst
@@ -17,6 +17,7 @@ New Features
 - Add support for environment variables in config (`#1037`_, pkvach)
 - Add Japanese localisation (`#1051`_, zurukumo)
 - Add Indonesian localisation (`#1088`_, suhindra)
+- Add Mistune as new Markdown rendering engine (`#1079`_, ritzmann)
 
 .. _#870: https://github.com/posativ/isso/pull/870
 .. _#966: https://github.com/posativ/isso/pull/966
@@ -28,6 +29,7 @@ New Features
 .. _#1037: https://github.com/isso-comments/isso/pull/1037
 .. _#1051: https://github.com/isso-comments/isso/pull/1051
 .. _#1088: https://github.com/isso-comments/isso/pull/1088
+.. _#1079: https://github.com/isso-comments/isso/pull/1079
 
 Breaking Changes
 ^^^^^^^^^^^^^^^^
@@ -38,6 +40,9 @@ Breaking Changes
   (tagged) version
 - Disavow IE10 compatibility: (Not-so) recent changes mean that Isso is no
   longer compatible (`#1022`, ix5)
+- The Markdown rendering engine Misaka has been deprecated. Please start using
+  Mistune, see :doc:`Migration from Misaka to Mistune <guides/mistune>` for
+  differences to Misaka.
 
 .. _#970: https://github.com/isso-comments/isso/pull/970
 .. _#1022: https://github.com/isso-comments/isso/pull/1022